+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * st-sidebar: Add functionality to close the expanded items when active item changes
 * st-foregrounds-notifications: Add property autoclose for notifications.
 
+**Fixed bugs:**
+
+* st-two-list-selection: Fix search functionality
+
+
 ## 11.1.0 (June 04, 2018)
 
 **New features:**